Update library to use Phaxio API to use v2.1
Phaxio has released a new API, so we need to update this client library to
support those changes. There are three major updates: a new endpoint,
removing parameters from the fax callback, and moving to HTTP Basic
Authorization.

1) We're going to use a new endpoint for the library (a different version
string). Only the tests needed to be rewritten for this change besides
changing the String in the Phaxio class.

2) The callback example project needed to remove `success`, `is_test`,
and `direction` from the expected parameters.

3) We needed to move to HTTP Basic Authorization. The RestClient
accomodated this change, but called it Authentication, so I renamed
the classes/methods from using the word authentication to authorization,
and then changed the Phaxio class to use basic auth on the RestClient.
This necessitated a bunch of changes to the tests that mocked the HTTP
endpoints which is the majority of the lines in this commit.

Finally, a couple of minor changes: a previous PR didn't space parameters
in method signatures and calls, so there are a few lines that just adding
spaces. Secondly, barcodes (an list of objects) and callerName (a String)
were added the Fax objects in the client and the example project.